Consolidate test structures and scripts
