+++ exray/package.json
@@ -11,8 +11,8 @@
     "types"
   ],
   "scripts": {
-    "build": "rm -rf dist && tsc",
-    "test": "tsc --outDir tmp --rootDir . test/index.ts && { node tmp/test/index.js ; rm -r tmp ; }",
+    "build": "rm -rf dist types && tsc && node dist/test.js && rm dist/test.js",
+    "test": "tsc --noEmit",
     "clean": "rm -rf node_modules package-lock.json dist types docs",
     "docs": "rm -rf docs && typedoc --includeVersion --disableSources --hideGenerator src",
     "prepack": "npm run build",
